Updated u32 to u8 on the struct to save some bits

+++ include/trainer_slide.h
@@ -6,8 +6,8 @@
 
 struct MessageStatus
 {
-    u32 messageInitalized[TRAINER_SLIDE_ARRAY_SIZE];
-    u32 messagePlayed[TRAINER_SLIDE_ARRAY_SIZE];
+    u8 messageInitalized[TRAINER_SLIDE_ARRAY_SIZE];
+    u8 messagePlayed[TRAINER_SLIDE_ARRAY_SIZE];
 };
 
 enum TrainerSlideTargets ShouldDoTrainerSlide(u32, enum TrainerSlideType);
